[US4256] feat(operator-sdk): Moving "cmd/*"  to "cmd/ndm_daemonset/" for operator-sdk.
