+++ cmd/ndm_daemonset/app/command/device-list.go
@@ -22,7 +22,7 @@ import (
 	"html/template"
 	"os"
 
-	"github.com/openebs/node-disk-manager/cmd/controller"
+	"github.com/openebs/node-disk-manager/cmd/ndm_daemonset/controller"
 	"github.com/spf13/cobra"
 )
