+++ cmd/ndm_daemonset/app/command/device.go
