+++ cmd/ndm_daemonset/controller/disk.go
