+++ cmd/ndm_daemonset/controller/diskstore.go
