+++ cmd/ndm_daemonset/controller/sparsefilegenerator.go
