+++ cmd/ndm_daemonset/filter/filter_test.go
@@ -20,7 +20,7 @@ import (
 	"sync"
 	"testing"
 
-	"github.com/openebs/node-disk-manager/cmd/controller"
+	"github.com/openebs/node-disk-manager/cmd/ndm_daemonset/controller"
 	"github.com/stretchr/testify/assert"
 )
