+++ cmd/ndm_daemonset/filter/osdiskexcludefilter.go
@@ -20,7 +20,7 @@ import (
 	"strings"
 
 	"github.com/golang/glog"
-	"github.com/openebs/node-disk-manager/cmd/controller"
+	"github.com/openebs/node-disk-manager/cmd/ndm_daemonset/controller"
 	"github.com/openebs/node-disk-manager/pkg/util"
 )
