+++ cmd/ndm_daemonset/filter/pathfilter.go
@@ -19,7 +19,7 @@ package filter
 import (
 	"strings"
 
-	"github.com/openebs/node-disk-manager/cmd/controller"
+	"github.com/openebs/node-disk-manager/cmd/ndm_daemonset/controller"
 	"github.com/openebs/node-disk-manager/pkg/util"
 )
