+++ cmd/ndm_daemonset/probe/probe.go
@@ -18,7 +18,7 @@ package probe
 
 import (
 	"github.com/golang/glog"
-	"github.com/openebs/node-disk-manager/cmd/controller"
+	"github.com/openebs/node-disk-manager/cmd/ndm_daemonset/controller"
 )
 
 const (
