+++ cmd/ndm_daemonset/probe/udevprobe.go
@@ -20,7 +20,7 @@ import (
 	"errors"
 
 	"github.com/golang/glog"
-	"github.com/openebs/node-disk-manager/cmd/controller"
+	"github.com/openebs/node-disk-manager/cmd/ndm_daemonset/controller"
 	libudevwrapper "github.com/openebs/node-disk-manager/pkg/udev"
 	"github.com/openebs/node-disk-manager/pkg/udevevent"
 	"github.com/openebs/node-disk-manager/pkg/util"
